Add PhotoComponent
디렉토리 경로 수정.
PhotoComponent 추가
